resume_analyzer/requirements.txt
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.2
cssselect2==0.7.0
cycler==0.12.1
cymem==2.0.11
Django==5.1.6
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
fonttools==4.56.0
html5lib==1.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
murmurhash==1.0.12
narwhals==1.28.0
nltk==3.9.1
numpy==2.2.3
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pdfminer.six==20240706
pillow==11.1.0
plotly==6.0.0
preshed==3.0.9
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pyparsing==3.2.1
pypdf==5.3.0
PyPDF2==3.0.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
svglib==1.5.1
thinc==8.3.4
threadpoolctl==3.5.0
tinycss2==1.4.0
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
uritools==4.0.3
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
webencodings==0.5.1
wrapt==1.17.2
xhtml2pdf==0.2.17